Use new built-in mapposition macro, as new attempt to fix #14
The implementation in c3cd245077db411d812498fd978c3aab7b1bd3ef could possibly be failing in some circumstances if the map instance is not initialised sufficiently quickly.
This moves the read of the hash directly into the code tree after the map object is known to exist, and therefore should be a more reliable method.